refactor(core): discovery + typegen behind the node subpath
Route discovery (`discoverRoutes`, `findDawnApp`, `assertDawnRoutesDir`) and
tool typegen (`extractToolSchemasForRoute`, `extractToolTypesForRoute`) read
the filesystem and load the TypeScript compiler, but were re-exported from
`@dawn-ai/core`'s `.` barrel. No package here sets `"sideEffects": false`, so
esbuild cannot drop them: every consumer of the barrel — including the
`@dawn-ai/cli/fetch` entry, which never CALLS any of them — carried
`node:fs`, `node:fs/promises`, `node:path`, `node:url`, `tsx` and `typescript`
as import edges.

ENTRY-POINT CHOICE: extended the EXISTING `"./node"` subpath (added in #389)
rather than adding `"./discovery"` and `"./build"` — fewer public surfaces,
and both concerns are node-only by nature. `src/node-marker-fs.ts` gains a
`src/node.ts` barrel above it; `"./node"` now points at `dist/node.js`.

Only the modules that actually import `node:` moved. The PURE members of
`discovery/` and `typegen/` (`toRouteSegments`, `isPrivateSegment`,
`isRouteGroupSegment`, `renderDawnTypes`, `renderRouteTypes`,
`renderStateTypes`, `renderToolTypes`) stay on the `.` barrel — the fetch path
calls `toRouteSegments`, so moving it would have ADDED edges.

`runtime-registry.ts` was the one fetch-path file importing `discoverRoutes`.
Split into a pure `runtime-registry-core.ts` (registry shape, static-modules
builder, manifest→registry mapping) and the node wrapper that keeps the
existing `createRuntimeRegistry(appRoot, modules?)` signature. The fetch core
now reaches discovery through the existing
`RuntimeBootFallbacks.discoverRouteManifest` seam, which also shares the
per-appRoot walk memo with route execution.

PUBLIC API SURFACE CHANGE (for the changeset): the five symbols above and the
`ExtractToolSchemasOptions` / `ExtractToolTypesOptions` types now ship from
`@dawn-ai/core/node`.

Ratchet: 23 -> 13 pinned upstream edges, 4 -> 1 LOADER_EDGES.
