+++ packages/cli/src/commands/typegen.ts
@@ -1,4 +1,4 @@
-import { discoverRoutes, findDawnApp } from "@dawn-ai/core"
+import { discoverRoutes, findDawnApp } from "@dawn-ai/core/node"
 import type { Command } from "commander"
 
 import { CliError, type CommandIo, formatErrorMessage, writeLine } from "../lib/output.js"
