+++ packages/cli/src/lib/runtime/resolve-identity.ts
@@ -1,6 +1,6 @@
 import { existsSync } from "node:fs"
 import { join } from "node:path"
-import { discoverRoutes } from "@dawn-ai/core"
+import { discoverRoutes } from "@dawn-ai/core/node"
 import { parseNamespace, routeNamespaceKey } from "@dawn-ai/memory/namespace"
 import { CliError, formatErrorMessage } from "../output.js"
 import { loadRouteMemory } from "./load-memory.js"
