+++ packages/cli/vitest.config.ts
@@ -8,7 +8,7 @@ const rootDir = dirname(fileURLToPath(import.meta.url))
 export default defineConfig({
   resolve: {
     alias: {
-      "@dawn-ai/core/node": resolve(rootDir, "../core/src/node-marker-fs.ts"),
+      "@dawn-ai/core/node": resolve(rootDir, "../core/src/node.ts"),
       "@dawn-ai/core": resolve(rootDir, "../core/src/index.ts"),
       "@dawn-ai/langchain": resolve(rootDir, "../langchain/src/index.ts"),
       "@dawn-ai/langgraph": resolve(rootDir, "../langgraph/src/index.ts"),
