+++ packages/core/package.json
@@ -26,8 +26,8 @@
       "default": "./dist/index.js"
     },
     "./node": {
-      "types": "./dist/node-marker-fs.d.ts",
-      "default": "./dist/node-marker-fs.js"
+      "types": "./dist/node.d.ts",
+      "default": "./dist/node.js"
     }
   },
   "publishConfig": {
