+++ packages/core/src/index.ts
@@ -53,8 +53,13 @@ export {
   seedDawnConfig,
 } from "./config.js"
 export { config } from "./config-helper.js"
-export { discoverRoutes } from "./discovery/discover-routes.js"
-export { assertDawnRoutesDir, findDawnApp } from "./discovery/find-dawn-app.js"
+// DELIBERATELY NOT HERE: `discoverRoutes`, `findDawnApp`,
+// `assertDawnRoutesDir`, `extractToolSchemasForRoute` and
+// `extractToolTypesForRoute` ship from "@dawn-ai/core/node". They read the
+// filesystem / load the TypeScript compiler, and a barrel re-export is an
+// import edge even for consumers that never call them. Their PURE siblings
+// below (route segments, the type renderers) stay on this barrel — the fetch
+// path calls `toRouteSegments`.
 export {
   isPrivateSegment,
   isRouteGroupSegment,
@@ -64,10 +69,6 @@ export type { ResolveStateFieldsOptions } from "./state/resolve-state-fields.js"
 export { resolveStateFields } from "./state/resolve-state-fields.js"
 export type { ScopeInput, ToolOrigin } from "./tool-scope.js"
 export { resolveToolScope, toolOrigin } from "./tool-scope.js"
-export type { ExtractToolSchemasOptions } from "./typegen/extract-tool-schema.js"
-export { extractToolSchemasForRoute } from "./typegen/extract-tool-schema.js"
-export type { ExtractToolTypesOptions } from "./typegen/extract-tool-types.js"
-export { extractToolTypesForRoute } from "./typegen/extract-tool-types.js"
 export {
   renderDawnTypes,
   renderRouteTypes,
