+++ test/runtime/run-tool-scope.test.ts
@@ -29,7 +29,7 @@ import {
   runTypegen,
   streamResolvedRoute,
 } from "@dawn-ai/cli/runtime"
-import { discoverRoutes } from "@dawn-ai/core"
+import { discoverRoutes } from "@dawn-ai/core/node"
 import { type Aimock, collectRunResult, createAimock, script } from "@dawn-ai/testing"
 import { afterAll, beforeAll, expect, it } from "vitest"
